Pandas
The result of his effort was Pandas, a high-performance, open-source Python library. The name Pandas is derived from the "panel data", a term for multidimensional data.
Pandas is not only used in data science but also in other fields such as advertising, economics, recommendation systems, neuroscience, and natural language processing.

Pandas is build around a two-dimensional data structure: the DataFrame. Think of a DataFrame as a table with columns and rows.